Split Bluetooth header into a few smaller ones
