+++ inc/Bluetooth.hpp
@@ -1,19 +1,5 @@
 #pragma once
-#include "esp_event.h"
-
-#include "esp_bt.h"
-#include "esp_bt_defs.h"
-#include "esp_bt_main.h"
-
-#include "esp_gap_ble_api.h"
-
-#include "esp_gatt_common_api.h"
-#include "esp_gatts_api.h"
-
-#include <vector>
-#include <map>
-#include <cstring>
-#include <functional>
+#include <debug.hpp>
 
 #define BLE_SERVICES_PRINTF
 #define BLE_CHARACTERISTICS_PRINTF
@@ -22,8 +8,6 @@
 #ifdef BLE_INPUT_PRINTF
     #define MAX_OUTPUT_AMOUNT 40
 #endif
-
-typedef __uint8_t byte;
 
 namespace Perm
 {
@@ -55,142 +39,3 @@ namespace Prop
         ExtProp     = (1 << 7),
     };
 }
-
-extern esp_gatt_if_t GATTinterface;
-
-class GATTinstance
-{
-    protected:
-        esp_bt_uuid_t UUID;
-        uint16_t Handler;
-    public:
-        GATTinstance(uint32_t ID)
-        {
-            UUID.len = ID & 0xFFFF0000 ? sizeof(uint32_t) : sizeof(uint16_t);
-            memcpy(&UUID.uuid, &ID, UUID.len);
-        }
-        GATTinstance(uint8_t ID[ESP_UUID_LEN_128])
-        {
-            UUID.len = ESP_UUID_LEN_128;
-            memcpy(&UUID.uuid, ID, UUID.len);
-        }
-        #if defined BLE_SERVICES_PRINTF || defined BLE_CHARACTERISTICS_PRINTF
-        void ConsoleInfoOut()
-        {
-            printf("UUID: ");
-            for(uint8_t i = 0; i < UUID.len; ++i)
-                printf("%02X", UUID.uuid.uuid128[i]);
-            printf(", Handler: %d\n", this->Handler);
-        }
-        #endif
-        
-        uint16_t getHandler() { return Handler; }
-        void setHandler(uint16_t Handler) { this->Handler = Handler; }
-};
-
-#define getCorrectSize(size) size > MTU ? MTU : size
-
-class Characteristic : public GATTinstance
-{
-    private:
-        typedef std::function<void(const Characteristic*, esp_ble_gatts_cb_param_t*)> ReadCallback;
-        typedef std::function<void(Characteristic*, const uint16_t, const void*)> WriteCallback;
-    
-        esp_attr_value_t  CharData;
-        const esp_gatt_perm_t Permition;
-        const esp_gatt_char_prop_t Property;
-
-        bool DataAllocatedInsideObject = false;
-        void* Data = nullptr;
-        size_t DataSize = 0;
-        
-        ReadCallback ReadHandler = [](const Characteristic* ch, esp_ble_gatts_cb_param_t *param){
-            ch->Responce(ch->getData(), ch->getDataSize(), param);
-        };
-        WriteCallback WriteHandler = [](Characteristic* ch, const uint16_t len, const void* value){
-            ch->setData(value, len);
-        };
-        static uint16_t MTU;
-    public:
-        Characteristic(uint32_t _UUID, esp_gatt_perm_t, esp_gatt_char_prop_t);
-        Characteristic(uint8_t _UUID[ESP_UUID_LEN_128], esp_gatt_perm_t, esp_gatt_char_prop_t);
-        
-        void setReadCallback(ReadCallback call) { ReadHandler = call; }
-        void setWriteCallback(WriteCallback call) { WriteHandler = call; }
-        void callReadCallback(esp_ble_gatts_cb_param_t *param);
-        void callWriteCallback(esp_ble_gatts_cb_param_t *param);
-
-        void setData(const void* Data, size_t DataSize);
-        void setDynamicData(void* Data, size_t DataSize);
-        void* getData() const { return Data; }
-        size_t getDataSize() const { return DataSize; }
-
-        
-        template<class type>
-        void Notify(const type data, size_t size = sizeof(type), uint16_t connID = 0) const
-        {
-            if(Property & Prop::Notify && GATTinterface)
-                esp_ble_gatts_send_indicate(GATTinterface, connID, Handler, getCorrectSize(size), (uint8_t*)&data, false);
-        }
-        void Notify(const void* Data, size_t DataSize, uint16_t connected_device_id = 0) const;
-        void Responce(const void* Data, size_t DataSize, esp_ble_gatts_cb_param_t* Param) const;
-
-        void AttachToService(uint16_t ServiceHandler);
-        static void setMTU(uint16_t);
-
-        ~Characteristic();
-};
-
-class Service : public GATTinstance
-{
-    private:
-        esp_gatt_srvc_id_t serviceID;
-        
-        const std::vector<Characteristic*> Characteristics;
-        const size_t CharacteristicsSize;
-    public:
-        const std::vector<Characteristic*> getCharacteristics() const { return Characteristics; }
-        size_t getCharacteristicsSize() const { return CharacteristicsSize; }
-
-        Service(const uint32_t UUID, const std::vector<Characteristic*> Characteristics);
-        Service(uint8_t UUID[ESP_UUID_LEN_128], const std::vector<Characteristic*> Characteristics);
-
-        void Start();
-        void Create();
-};
-
-typedef void GATTScallbackType(esp_ble_gatts_cb_param_t*);
-typedef void GAPcallbackType(esp_ble_gap_cb_param_t*);
-class ServerDevice
-{
-    private:
-        static const uint8_t MaxEventNumber = 24;
-
-        static const char* Name;
-        static esp_ble_adv_params_t AdvertisingParameters;
-        static esp_ble_adv_data_t AdvertisingData;
-        static esp_ble_adv_data_t ScanResponceData;
-
-        static std::vector<Service*> Services;
-
-        static GATTScallbackType* DeviceCallbacks[MaxEventNumber];
-        static GAPcallbackType* GAPcalls[ESP_GAP_BLE_EVT_MAX];
-        static void HandleGATTSevents(esp_gatts_cb_event_t event, esp_gatt_if_t gatts_if, esp_ble_gatts_cb_param_t *param);
-        static void HandleGAPevents(esp_gap_ble_cb_event_t event, esp_ble_gap_cb_param_t *param);
-
-        static void Start();
-
-        static uint8_t serviceCounter;
-        static uint8_t charCounter;
-        static bool isEnabled;
-    public:
-        ServerDevice();
-        ServerDevice(const char* Name, std::initializer_list<Service*> Services);
-
-        void setGATTSevent(esp_gatts_cb_event_t Event, GATTScallbackType* Callback);
-        void setGAPevent(esp_gap_ble_cb_event_t Event, GAPcallbackType* call);
-
-        static bool Status() { return isEnabled; }
-        static void Enable();
-        static void Disable();
-};
