+++ src/Bluetooth.ServerDevice.cpp
@@ -1,4 +1,4 @@
-#include <Bluetooth.hpp>
+#include <Bluetooth.Server.hpp>
 
 ServerDevice::ServerDevice() {}
 bool ServerDevice::isEnabled = false;
